Migrate WebRTC Video room action
