+++ apps/meteor/app/webrtc/client/index.ts
@@ -1,5 +1,4 @@
 import './adapter';
-import './tabBar';
 import './actionLink';
 
 export * from './WebRTCClass';
